d3heatmap {d3heatmap}	R Documentation

D3 Heatmap widget

Description

Creates a D3.js-based heatmap widget.

Usage

d3heatmap(x, Rowv = TRUE, Colv = if (symm) "Rowv" else TRUE,
  distfun = dist, hclustfun = hclust, dendrogram = c("both", "row",
  "column", "none"), reorderfun = function(d, w) reorder(d, w), k_row, k_col,
  symm = FALSE, revC, scale = c("none", "row", "column"), na.rm = TRUE,
  labRow = rownames(x), labCol = colnames(x), cexRow, cexCol, digits = 3L,
  cellnote, cellnote_scale = FALSE, theme = NULL, colors = "RdYlBu",
  width = NULL, height = NULL, xaxis_height = 80, yaxis_width = 120,
  xaxis_font_size = NULL, yaxis_font_size = NULL, brush_color = "#0000FF",
  show_grid = TRUE, anim_duration = 500, ...)

Arguments

x
A numeric matrix Defaults to TRUE unless x contains any NAs.

Rowv
determines if and how the row dendrogram should be reordered. By default, it is TRUE, which implies dendrogram is computed and reordered based on row means. If NULL or FALSE, then no dendrogram is computed and no reordering is done. If a dendrogram, then it is used "as-is", ie without any reordering. If a vector of integers, then dendrogram is computed and reordered based on the order of the vector.

Colv
determines if and how the column dendrogram should be reordered. Has the options as the Rowv argument above and additionally when x is a square matrix, Colv = "Rowv" means that columns should be treated identically to the rows.

distfun
function used to compute the distance (dissimilarity) between both rows and columns. Defaults to dist.

hclustfun
function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms. Defaults to hclust.

dendrogram
character string indicating whether to draw 'none', 'row', 'column' or 'both' dendrograms. Defaults to 'both'. However, if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a warning is issued and Rowv (or Colv) arguments are honoured.

reorderfun
function(d, w) of dendrogram and weights for reordering the row and column dendrograms. The default uses statsreorder.dendrogram

k_row
an integer scalar with the desired number of groups by which to color the dendrogram's branches in the rows (uses color_branches)

k_col
an integer scalar with the desired number of groups by which to color the dendrogram's branches in the columns (uses color_branches)

symm
logical indicating if x should be treated symmetrically; can only be true when x is a square matrix.

revC
logical indicating if the column order should be reversed for plotting. Default (when missing) - is FALSE, unless symm is TRUE. This is useful for cor matrix.

scale
character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".

na.rm
logical indicating whether NA's should be removed.

labRow
character vectors with row labels to use (from top to bottom); default to rownames(x).

labCol
character vectors with column labels to use (from left to right); default to colnames(x).

cexRow
positive numbers. If not missing, it will override xaxis_font_size and will give it a value cexRow*14

cexCol
positive numbers. If not missing, it will override yaxis_font_size and will give it a value cexCol*14

digits
integer indicating the number of decimal places to be used by round for 'label'.

cellnote
(optional) matrix of the same dimensions as x that has the human-readable version of each value, for displaying to the user on hover. If NULL, then x will be coerced using as.character. If missing, it will use x, after rounding it based on the digits parameter.

cellnote_scale
logical (default is FALSE). IF cellnote is missing and x is used, should cellnote be scaled if x is also scaled?

theme
A custom CSS theme to use. Currently the only valid values are "" and "dark". "dark" is primarily intended for standalone visualizations, not R Markdown or Shiny.

colors
Either a colorbrewer2.org palette name (e.g. "YlOrRd" or "Blues"), or a vector of colors to interpolate in hexadecimal "#RRGGBB" format, or a color interpolation function like colorRamp.

width
Width in pixels (optional, defaults to automatic sizing).

height
Height in pixels (optional, defaults to automatic sizing).

xaxis_height
Size of axes, in pixels.

yaxis_width
Size of axes, in pixels.

xaxis_font_size
Font size of axis labels, as a CSS size (e.g. "14px" or "12pt").

yaxis_font_size
Font size of axis labels, as a CSS size (e.g. "14px" or "12pt").

brush_color
The base color to be used for the brush. The brush will be filled with a low-opacity version of this color. "#RRGGBB" format expected.

show_grid
TRUE to show gridlines, FALSE to hide them, or a numeric value to specify the gridline thickness in pixels (can be a non-integer).

anim_duration
Number of milliseconds to animate zooming in and out. For large x it may help performance to set this value to 0.

...
currently ignored

Source

The interface was designed based on heatmap and heatmap.2

See Also

heatmap, heatmap.2

Examples

library(d3heatmap)
d3heatmap(mtcars, scale = "column", colors = "Blues")

[Package d3heatmap version 0.6.1.2 Index]