Modify cell options

Provides options for modifying the cells and the display of their values

hmCells(
  d3heatmap,
  digits,
  cellnotes,
  font.size,
  color,
  scale,
  row.label,
  col.label,
  value.label,
  brush.color,
  print
)

Arguments

d3heatmap
Required A valid d3heatmap object

digits
integer (Optional) indicating the number of decimal places to be used by round for 'label'.

cellnotes
numeric (Optional) matrix of the same dimensions as x that has the human-readable version of each value, for displaying to the user on hover. If NULL, then x will be coerced using as.character. If missing, it will use x, after rounding it based on the digits parameter.

font.size
numeric the pixel size of printed cell

color
character name or hex specifying the color of the values printed inside the cells

scale
logical (default is FALSE). If cellnote is missing and x is used, should cellnotes be scaled if x is also scaled?

row.label
character (Optional) Label to display next to the row value when the user hovers over the cell. If not specified, tries to match the xaxis_title; if no axis title, defaults to "Row".

col.label
character (Optional) Label to display next to the column value when the user hovers over the cell If not specified, tries to match the yaxis_title; if no axis title, defaults to "Col".

value.label
character (Optional) Label to display next to the cell value when the user hovers over the cell. Defaults to "Value".

brush.color
The base color to be used for the brush. The brush will be filled with a low-opacity version of this color. "#RRGGBB" format expected.

print
logical Show the values inside the cells. Defaults to FALSE.

Source

The interface was inspired by dygraphs

Value

Modified d3heatmap object

See also

Examples

if (FALSE) {

d3heatmap(mtcars, scale = "column", col = "Blues") %>%
  hmCells(digits = 0L, print = TRUE)

}